requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
fastapi
uvicorn[standard]
httpx
python-dotenv
